Create the necessary infrastructure for users CRUD.
Using the examples from this lesson, secure your blog application such that you have to log in to create a post.

Test this functionality, ensure that you get redirected to the login page if you try to create a post without being logged in.

When a post is created, assign the logged in user as the owner of that post.